Refactor load to user FreqmanDB
