+++ firmware/application/apps/ui_freqman.cpp
@@ -24,7 +24,6 @@
 #include "ui_freqman.hpp"
 
 #include "event_m0.hpp"
-#include "freqman.hpp"
 #include "portapack.hpp"
 #include "rtc_time.hpp"
 #include "tone_key.hpp"
@@ -35,11 +34,33 @@
 #include <memory>
 
 using namespace portapack;
+using namespace ui;
 namespace fs = std::filesystem;
 
 // TODO: Clean up after moving to better lookup tables.
-extern ui::OptionsField::options_t freqman_bandwidths[4];
-extern ui::OptionsField::options_t freqman_steps;
+using options_t = OptionsField::options_t;
+extern options_t freqman_modulations;
+extern options_t freqman_bandwidths[4];
+extern options_t freqman_steps;
+extern options_t freqman_steps_short;
+
+/* Set options. */
+void freqman_set_modulation_option(OptionsField& option) {
+    option.set_options(freqman_modulations);
+}
+
+void freqman_set_bandwidth_option(freqman_index_t modulation, OptionsField& option) {
+    if (is_valid(modulation))
+        option.set_options(freqman_bandwidths[modulation]);
+}
+
+void freqman_set_step_option(OptionsField& option) {
+    option.set_options(freqman_steps);
+}
+
+void freqman_set_step_option_short(OptionsField& option) {
+    option.set_options(freqman_steps_short);
+}
 
 namespace ui {
